mito-ai: let exceptions bubble
