+++ mito-ai/mito_ai/utils/anthropic_utils.py
@@ -84,7 +84,7 @@ async def get_anthropic_completion_from_mito_server(
         model, max_tokens, temperature, system, messages, message_type, tools, tool_choice, None
     )
     
-    return await get_response_from_mito_server(MITO_ANTHROPIC_URL, headers, data, timeout, max_retries)
+    return await get_response_from_mito_server(MITO_ANTHROPIC_URL, headers, data, timeout, max_retries, message_type)
 
 async def stream_anthropic_completion_from_mito_server(
     model: Union[str, None],
