+++ mito-ai/mito_ai/utils/open_ai_utils.py
@@ -92,7 +92,7 @@ async def get_ai_completion_from_mito_server(
         message_type
     )
     
-    return await get_response_from_mito_server(MITO_OPENAI_URL, headers, data, timeout, max_retries)
+    return await get_response_from_mito_server(MITO_OPENAI_URL, headers, data, timeout, max_retries, message_type)
 
 async def stream_ai_completion_from_mito_server(
     last_message_content: Union[str, None],
